minor fix added
