+++ src/Common/CustomTagSelector/index.ts
@@ -1,3 +1,6 @@
+export {default as PropagateTagInfo} from './PropagateTagInfo'
+export * from './TagDetails'
 export * from './TagLabelSelect'
 export * from './TagLabelValueSelector'
 export * from './ResizableTagTextArea'
+export * from './Types'
